+++ src/routes/engineering/ie.ts
@@ -1,18 +1,21 @@
 // filename must equal to first level of url domain.
-// e.g. mse.snu.ac.kr -> mse.ts
+// e.g. ie.snu.ac.kr -> ie.ts
 
+import assert from 'assert';
+import * as Apify from 'apify';
 import { CheerioHandlePageInputs } from 'apify/types/crawlers/cheerio_crawler';
 import { RequestQueue } from 'apify';
 import { load } from 'cheerio';
+import { Connection } from 'typeorm';
 import { URL } from 'url';
-import { Notice, File } from '../../server/src/notice/notice.entity.js';
-import { SiteData } from '../types/custom-types';
-import { absoluteLink, getOrCreate, getOrCreateTags, saveNotice } from '../utils';
-import { strptime } from '../micro-strptime';
-import { CategoryCrawler } from '../classes/categoryCrawler.js';
-import { ENGINEERING } from '../constants';
+import { Notice, File } from '../../../server/src/notice/notice.entity.js';
+import { SiteData } from '../../types/custom-types';
+import { absoluteLink, getOrCreate, getOrCreateTags, saveNotice } from '../../utils';
+import { strptime } from '../../micro-strptime';
+import { CategoryCrawler } from '../../classes/categoryCrawler.js';
+import { ENGINEERING } from '../../constants';
 
-class MSECrawler extends CategoryCrawler {
+class IECrawler extends CategoryCrawler {
     handlePage = async (context: CheerioHandlePageInputs): Promise<void> => {
         const { request, $ } = context;
         const { url } = request;
@@ -23,19 +26,28 @@ class MSECrawler extends CategoryCrawler {
             // creation order
             // dept -> notice -> file
             //                -> tag -> notice_tag
-
+            $('img').each((index, element) => {
+                const imgSrc = $(element).attr('src');
+                $(element).attr('src', absoluteLink(imgSrc, this.baseUrl) ?? '');
+            });
             const notice = await getOrCreate(Notice, { link: url }, false);
 
             notice.department = siteData.department;
-            const title = $('td[id="fn"]').text();
+            const title = $('div[property="dc:title"]').children('h2').text();
             notice.title = title;
-            const contentElement = $('td[class="s_default_view_body_2"]').find('td');
-            contentElement.find('div[class="mnSns"]').remove();
+            const contentElement = $('div[property="content:encoded"]');
+
             const content = load(contentElement.html() ?? '', { decodeEntities: false })('body').html() ?? '';
             // ^ encode non-unicode letters with utf-8 instead of HTML encoding
             notice.content = content;
             notice.preview = contentElement.text().substring(0, 1000).trim(); // texts are automatically utf-8 encoded
-            notice.createdAt = strptime(siteData.dateString, '%Y-%m-%d');
+            const fullDateString: string = $('div.field-name-post-date').find('div.field-item').text().trim();
+
+            try {
+                notice.createdAt = strptime(fullDateString, '%Y-%m-%d %H:%M:%S');
+            } catch {
+                notice.createdAt = strptime(siteData.dateString, '%Y-%m-%d');
+            }
 
             notice.isPinned = siteData.isPinned;
             notice.link = url;
@@ -43,15 +55,17 @@ class MSECrawler extends CategoryCrawler {
             await saveNotice(notice);
 
             const files: File[] = [];
-            $('#boardSkin_s_default_view > tbody > tr:nth-child(1) > td a').each((index, element) => {
-                const fileUrl = $(element).attr('href');
-                if (fileUrl) {
-                    const file = new File();
-                    file.name = $(element).text().trim();
-                    file.link = absoluteLink(fileUrl, this.baseUrl) ?? '';
-                    files.push(file);
-                }
-            });
+            $('div.field-name-field-attachment')
+                .find('span.file')
+                .each((index, element) => {
+                    const fileUrl = $(element).children('a').attr('href');
+                    if (fileUrl) {
+                        const file = new File();
+                        file.name = $(element).children('a').text().trim();
+                        file.link = fileUrl;
+                        files.push(file);
+                    }
+                });
             await Promise.all(
                 // using Promise.all in order to ensure full execution
                 files.map(async (file) => {
@@ -60,14 +74,8 @@ class MSECrawler extends CategoryCrawler {
                 }),
             );
 
-            const category = new URL(url).searchParams.get('category') ?? '64';
-            const tags: string[] = [];
-            if (siteData.tag) {
-                tags.push(siteData.tag);
-            }
-            if (!tags.includes(this.categoryTags[category])) {
-                tags.push(this.categoryTags[category]);
-            }
+            const category = new URL(url).pathname.split('/')[3]; // url.replace(BaseUrl, '').split('?')[0];
+            const tags = [this.categoryTags[category]];
             await getOrCreateTags(tags, notice, siteData.department);
         }
     };
@@ -80,25 +88,16 @@ class MSECrawler extends CategoryCrawler {
 
         if ($) {
             $('tbody tr').each((index, element) => {
-                const titleElement = $(element).find('td:nth-child(2) a');
-                const noticeIdxRe = /viewData\('([0-9]+)'\)/;
-                const noticeIdx = titleElement.attr('onclick')?.match(noticeIdxRe);
-                if (!noticeIdx) return;
-                const pageUrl = new URL(url);
-                pageUrl.searchParams.set('mode', 'view');
-                pageUrl.searchParams.set('board_num', noticeIdx[1]);
-                pageUrl.searchParams.delete('page');
-                const link = pageUrl.href;
-
-                const tag = $(element).find('td:nth-child(1)').text();
+                const titleElement = $(element).children('td.views-field-title-field').children('a');
+                const link = absoluteLink(titleElement.attr('href'), request.loadedUrl);
                 if (link === undefined) return;
-                const dateString = $(element).children('td').slice(3, 4).text().trim();
+                const dateString = $(element).find('td.views-field-created').text().trim();
+
                 const newSiteData: SiteData = {
                     department: siteData.department,
                     isPinned: false,
                     isList: false,
                     dateString,
-                    tag,
                 };
                 this.log.info('Enqueueing', { link });
                 requestQueue.addRequest({
@@ -107,7 +106,7 @@ class MSECrawler extends CategoryCrawler {
                 });
             });
 
-            const endElement = $('tfoot tr td a').last().attr('href');
+            const endElement = $('ul.pagination').children('li.pager-last').children('a').attr('href');
             const endUrl = absoluteLink(endElement, request.loadedUrl);
             if (!endUrl) return;
             const endUrlInstance = new URL(endUrl);
@@ -135,14 +134,20 @@ class MSECrawler extends CategoryCrawler {
     };
 }
 
-export const mse = new MSECrawler({
-    departmentName: '재료공학부',
-    departmentCode: 'mse',
-    baseUrl: 'https://mse.snu.ac.kr/sub.php?code=notice&category=',
+export const ie = new IECrawler({
+    departmentName: '산업공학과',
+    departmentCode: 'ie',
+    baseUrl: 'http://ie.snu.ac.kr/ko/board/',
     departmentCollege: ENGINEERING,
     categoryTags: {
-        1: '학부',
-        2: '대학원',
-        64: '전체',
+        2: '학과 주요뉴스',
+        3: '학과 행사',
+        4: '자료실',
+        5: '기타사항',
+        6: '취업',
+        7: '학부',
+        8: '대학원',
+        //  14:'역대 이중한 상 수상자',
+        15: '장학금',
     },
 });
