+++ src/routes/engineering/mse.ts
@@ -1,29 +1,18 @@
 // filename must equal to first level of url domain.
-// e.g. ee.snu.ac.kr -> ee.ts
+// e.g. mse.snu.ac.kr -> mse.ts
 
-import assert from 'assert';
-import * as Apify from 'apify';
 import { CheerioHandlePageInputs } from 'apify/types/crawlers/cheerio_crawler';
 import { RequestQueue } from 'apify';
 import { load } from 'cheerio';
-import { Connection } from 'typeorm';
 import { URL } from 'url';
-import { Notice, File } from '../../server/src/notice/notice.entity.js';
-import { CategoryCrawlerInit, CategoryTag, SiteData } from '../types/custom-types';
-import { absoluteLink, getOrCreate, getOrCreateTags, runCrawler, saveNotice } from '../utils';
-import { Department } from '../../server/src/department/department.entity';
-import { strptime } from '../micro-strptime';
-import { CategoryCrawler } from '../classes/categoryCrawler';
-import { ENGINEERING } from '../constants';
+import { Notice, File } from '../../../server/src/notice/notice.entity.js';
+import { SiteData } from '../../types/custom-types';
+import { absoluteLink, getOrCreate, getOrCreateTags, saveNotice } from '../../utils';
+import { strptime } from '../../micro-strptime';
+import { CategoryCrawler } from '../../classes/categoryCrawler.js';
+import { ENGINEERING } from '../../constants';
 
-class EECrawler extends CategoryCrawler {
-    private readonly excludeTags: string[];
-
-    constructor(initData: CategoryCrawlerInit) {
-        super(initData);
-        this.excludeTags = ['sugang', 'yonhapai'];
-    }
-
+class MSECrawler extends CategoryCrawler {
     handlePage = async (context: CheerioHandlePageInputs): Promise<void> => {
         const { request, $ } = context;
         const { url } = request;
@@ -34,35 +23,32 @@ class EECrawler extends CategoryCrawler {
             // creation order
             // dept -> notice -> file
             //                -> tag -> notice_tag
-            $('img').each((index, element) => {
-                const imgSrc = $(element).attr('src');
-                $(element).attr('src', absoluteLink(imgSrc, this.baseUrl) ?? '');
-            });
+
             const notice = await getOrCreate(Notice, { link: url }, false);
 
             notice.department = siteData.department;
-            const title = $('div#bbs-view-wrap').children('h1').text();
+            const title = $('td[id="fn"]').text();
             notice.title = title;
-            const contentElement = $('div.cnt');
-
+            const contentElement = $('td[class="s_default_view_body_2"]').find('td');
+            contentElement.find('div[class="mnSns"]').remove();
             const content = load(contentElement.html() ?? '', { decodeEntities: false })('body').html() ?? '';
             // ^ encode non-unicode letters with utf-8 instead of HTML encoding
             notice.content = content;
             notice.preview = contentElement.text().substring(0, 1000).trim(); // texts are automatically utf-8 encoded
-
             notice.createdAt = strptime(siteData.dateString, '%Y-%m-%d');
+
             notice.isPinned = siteData.isPinned;
             notice.link = url;
 
             await saveNotice(notice);
 
             const files: File[] = [];
-            $('div.att-file ul li div').each((index, element) => {
-                const fileUrl = $(element).children('a').attr('href');
+            $('#boardSkin_s_default_view > tbody > tr:nth-child(1) > td a').each((index, element) => {
+                const fileUrl = $(element).attr('href');
                 if (fileUrl) {
                     const file = new File();
                     file.name = $(element).text().trim();
-                    file.link = url;
+                    file.link = absoluteLink(fileUrl, this.baseUrl) ?? '';
                     files.push(file);
                 }
             });
@@ -74,10 +60,13 @@ class EECrawler extends CategoryCrawler {
                 }),
             );
 
-            const category = url.replace(this.baseUrl, '').split('?')[0];
-            const tags = [this.categoryTags[category]];
-            if (!this.excludeTags.includes(category) && title.startsWith('[')) {
-                tags.push(title.slice(1, title.indexOf(']')).trim());
+            const category = new URL(url).searchParams.get('category') ?? '64';
+            const tags: string[] = [];
+            if (siteData.tag) {
+                tags.push(siteData.tag);
+            }
+            if (!tags.includes(this.categoryTags[category])) {
+                tags.push(this.categoryTags[category]);
             }
             await getOrCreateTags(tags, notice, siteData.department);
         }
@@ -90,21 +79,26 @@ class EECrawler extends CategoryCrawler {
         this.log.info('Page opened.', { url });
 
         if ($) {
-            $('div.bbs-blogstyle ul li').each((index, element) => {
-                const titleElement = $(element).children('a').first();
-                // const title = titleElement.children('strong').first().text();
-                let link = absoluteLink(titleElement.attr('href'), request.loadedUrl);
-                if (link === undefined) return;
-                const pageUrl = new URL(link);
+            $('tbody tr').each((index, element) => {
+                const titleElement = $(element).find('td:nth-child(2) a');
+                const noticeIdxRe = /viewData\('([0-9]+)'\)/;
+                const noticeIdx = titleElement.attr('onclick')?.match(noticeIdxRe);
+                if (!noticeIdx) return;
+                const pageUrl = new URL(url);
+                pageUrl.searchParams.set('mode', 'view');
+                pageUrl.searchParams.set('board_num', noticeIdx[1]);
                 pageUrl.searchParams.delete('page');
-                link = pageUrl.href;
-                const dateString = $(element).find('p.date span').text().split('l')[1].trim();
+                const link = pageUrl.href;
 
+                const tag = $(element).find('td:nth-child(1)').text();
+                if (link === undefined) return;
+                const dateString = $(element).children('td').slice(3, 4).text().trim();
                 const newSiteData: SiteData = {
                     department: siteData.department,
                     isPinned: false,
                     isList: false,
                     dateString,
+                    tag,
                 };
                 this.log.info('Enqueueing', { link });
                 requestQueue.addRequest({
@@ -113,9 +107,8 @@ class EECrawler extends CategoryCrawler {
                 });
             });
 
-            const endElement = $('div.pagination-01').children('a.direction.next').last().attr('href');
+            const endElement = $('tfoot tr td a').last().attr('href');
             const endUrl = absoluteLink(endElement, request.loadedUrl);
-
             if (!endUrl) return;
             const endUrlInstance = new URL(endUrl);
             const urlInstance = new URL(url);
@@ -142,18 +135,14 @@ class EECrawler extends CategoryCrawler {
     };
 }
 
-export const ee = new EECrawler({
-    departmentName: '전기정보공학부',
-    departmentCode: 'ee',
-    baseUrl: 'https://ee.snu.ac.kr/community/notice/',
+export const mse = new MSECrawler({
+    departmentName: '재료공학부',
+    departmentCode: 'mse',
+    baseUrl: 'https://mse.snu.ac.kr/sub.php?code=notice&category=',
     departmentCollege: ENGINEERING,
     categoryTags: {
-        academic: '학사',
-        scholarship: '장학',
-        admissions: '입시&기타',
-        campuslife: '대학생활',
-        jobs: '취업&전문연',
-        sugang: '수강',
-        yonhapai: '인공지능',
+        1: '학부',
+        2: '대학원',
+        64: '전체',
     },
 });
