+++ src/routes/engineering/ship.ts
@@ -1,14 +1,17 @@
-import { CheerioHandlePageInputs } from 'apify/types/crawlers/cheerio_crawler';
-import { load } from 'cheerio';
-import { RequestQueue } from 'apify';
-import { CategoryCrawler } from '../classes/categoryCrawler';
-import { SCIENCE } from '../constants';
-import { SiteData } from '../types/custom-types';
-import { absoluteLink, getOrCreate, getOrCreateTags, saveNotice } from '../utils';
-import { File, Notice } from '../../server/src/notice/notice.entity';
-import { strptime } from '../micro-strptime';
+// filename must equal to first level of url domain.
+// e.g. cse.snu.ac.kr -> cse.ts
 
-class MathCrawler extends CategoryCrawler {
+import { CheerioHandlePageInputs } from 'apify/types/crawlers/cheerio_crawler';
+import { RequestQueue } from 'apify';
+import { load } from 'cheerio';
+import { Notice, File } from '../../../server/src/notice/notice.entity.js';
+import { SiteData } from '../../types/custom-types';
+import { absoluteLink, getOrCreate, getOrCreateTags, runCrawler, saveNotice, removeUrlPageParam } from '../../utils';
+import { strptime } from '../../micro-strptime';
+import { Crawler } from '../../classes/crawler';
+import { ENGINEERING } from '../../constants';
+
+class ShipCrawler extends Crawler {
     handlePage = async (context: CheerioHandlePageInputs): Promise<void> => {
         const { request, $ } = context;
         const { url } = request;
@@ -24,22 +27,20 @@ class MathCrawler extends CategoryCrawler {
             const notice = await getOrCreate(Notice, { link: url }, false);
 
             notice.department = siteData.department;
-            notice.title = $('div.titleAndUser div.title h2 a, div.titleAndUser div.title h1 a ').text().trim();
-            const contentElement = $('div.contentBody div.xe_content');
-            let content;
-            if ($('div.contentBody div.xe_content div.document_popup_menu').length) {
-                contentElement.find('div.document_popup_menu').remove();
-                content = contentElement.html() ?? '';
-            } else {
-                content = contentElement.html() ?? '';
-            }
+            notice.title = $('div.title h1 a').text().trim();
+            const contentElement = $('div.xe_content');
+            let content = contentElement.html() ?? '';
             content = load(content, { decodeEntities: false })('body').html() ?? '';
             // ^ encode non-unicode letters with utf-8 instead of HTML encoding
             notice.content = content;
             notice.preview = contentElement.text().substring(0, 1000).trim(); // texts are automatically utf-8 encoded
-            const fullDateString: string = $('div.dateAndCount div.date').text().trim();
-            // example: '2021-02-26 11:34:01'
-            notice.createdAt = strptime(fullDateString, '%Y.%m.%d %H:%M:%S');
+
+            try {
+                const fullDateString: string = $('div.date').text().trim();
+                notice.createdAt = strptime(fullDateString, '%Y.%m.%d %H:%M:%S');
+            } catch {
+                notice.createdAt = strptime(siteData.dateString, '%Y-%m-%d');
+            }
 
             notice.isPinned = siteData.isPinned;
             notice.link = url;
@@ -52,7 +53,7 @@ class MathCrawler extends CategoryCrawler {
                 if (fileUrl) {
                     const file = new File();
                     file.name = $(element).text().trim();
-                    file.link = absoluteLink(fileUrl, this.baseUrl) ?? '';
+                    file.link = url;
                     files.push(file);
                 }
             });
@@ -66,11 +67,8 @@ class MathCrawler extends CategoryCrawler {
             );
 
             const tags: string[] = [];
-            const urlInstance = new URL(request.loadedUrl);
-            const category = urlInstance.searchParams.get('mid');
-            if (category && this.categoryTags[category]) {
-                tags.push(this.categoryTags[category]);
-            }
+            const category = $('div.category a').text().trim();
+            tags.push(category);
             await getOrCreateTags(tags, notice, siteData.department);
         }
     };
@@ -81,21 +79,26 @@ class MathCrawler extends CategoryCrawler {
         const siteData = <SiteData>request.userData;
         this.log.info('Page opened.', { url });
         if ($) {
-            $('table.boardList tbody tr').each((index, element) => {
+            const urlInstance = new URL(url);
+            const page: number = +(urlInstance.searchParams.get('page') ?? 1);
+            // example:  /ko/board/Scholarship/page/2 => ['', 'ko', 'board', 'Scholarship','page','2']
+
+            $('table.boardList tr').each((index, element) => {
+                const isPinned = $(element).children('td.notice').length !== 0;
+                if (page > 1 && isPinned) return;
+
                 const titleElement = $(element).find('td.title a').first();
                 // const title = titleElement.text();
-
-                let link = absoluteLink(titleElement.attr('href'), request.loadedUrl);
+                const link = removeUrlPageParam(absoluteLink(titleElement.attr('href'), request.loadedUrl));
                 if (link === undefined) return;
-                const pageUrl = new URL(link);
-                pageUrl.searchParams.delete('page');
-                link = pageUrl.href;
+
+                const dateString = $(element).children('td.date').text();
 
                 const newSiteData: SiteData = {
                     department: siteData.department,
-                    isPinned: false,
+                    isPinned,
                     isList: false,
-                    dateString: '',
+                    dateString,
                 };
                 this.log.info('Enqueueing', { link });
                 requestQueue.addRequest({
@@ -104,29 +107,26 @@ class MathCrawler extends CategoryCrawler {
                 });
             });
 
-            const endElement = $('div.pagination.a1 a.nextEnd');
-            const endUrl = absoluteLink(endElement.attr('href'), request.loadedUrl);
-            const endUrlInstance = new URL(endUrl ?? '');
-            const urlInstance = new URL(request.loadedUrl);
-            if (!endUrlInstance || !urlInstance) return;
+            const endUrlInstance = new URL(absoluteLink($('div.pagination.a1 a.nextEnd').attr('href'), url) ?? '');
+            if (!endUrlInstance) return;
 
             const endPage = +(endUrlInstance.searchParams.get('page') ?? 1);
-            const page = +(urlInstance.searchParams.get('page') ?? 1);
-
+            // +lastNoticeId === 1  <==> loaded page is the last page
             if (page < endPage) {
-                const nextUrlInstance = new URL(urlInstance.href);
-                nextUrlInstance.searchParams.set('page', (page + 1).toString());
-                const nextList = nextUrlInstance.href;
+                const nextListInstance = new URL(urlInstance.href);
+                nextListInstance.searchParams.set('page', (page + 1).toString());
 
-                this.log.info('Enqueueing list', { nextList });
+                this.log.info('Enqueueing list', { nextList: nextListInstance.href });
+
                 const nextListSiteData: SiteData = {
                     department: siteData.department,
                     isPinned: false,
                     isList: true,
                     dateString: '',
                 };
-                await this.addVaryingRequest(requestQueue, {
-                    url: nextList,
+
+                await requestQueue.addRequest({
+                    url: nextListInstance.href,
                     userData: nextListSiteData,
                 });
             }
@@ -134,13 +134,9 @@ class MathCrawler extends CategoryCrawler {
     };
 }
 
-export const math = new MathCrawler({
-    departmentName: '수리과학부',
-    departmentCode: 'math', // this value must be equal to the filename
-    departmentCollege: SCIENCE,
-    baseUrl: 'http://www.math.snu.ac.kr/board/index.php?mid=',
-    categoryTags: {
-        notice: '공지사항',
-        employment: '행사/취업',
-    },
+export const ship = new ShipCrawler({
+    departmentName: '조선해양공학과',
+    departmentCode: 'ship', // this value must be equal to the filename,
+    baseUrl: 'http://ship.snu.ac.kr/index.php?mid=Notice',
+    departmentCollege: ENGINEERING,
 });
