+++ src/routes/routeList.ts
@@ -1,26 +1,28 @@
 import { Connection } from 'typeorm';
-import { cse } from './cse';
-import { ere } from './ere';
-import { ee } from './ee';
-import { cbe } from './cbe';
-import { me } from './me';
-import { ie } from './ie';
-import { math } from './math';
-import { ship } from './ship';
-import { cee } from './cee';
-import { architecture } from './architecture';
-import { mse } from './mse';
+import { cse } from './engineering/cse';
+import { ere } from './engineering/ere';
+import { ee } from './engineering/ee';
+import { cbe } from './engineering/cbe';
+import { me } from './engineering/me';
+import { ie } from './engineering/ie';
+import { math } from './science/math';
+import { ship } from './engineering/ship';
+import { cee } from './engineering/cee';
+import { architecture } from './engineering/architecture';
+import { mse } from './engineering/mse';
+import { stat } from './science/stat';
 
 export const routeList: ((connection: Connection) => Promise<void>)[] = [
-    cse.startCrawl,
-    ee.startCrawl,
-    cbe.startCrawl,
-    ere.startCrawl,
-    me.startCrawl,
-    ie.startCrawl,
-    ship.startCrawl,
-    cee.startCrawl,
-    architecture.startCrawl,
-    mse.startCrawl,
-    math.startCrawl,
+    // cse.startCrawl,
+    // ee.startCrawl,
+    // cbe.startCrawl,
+    // ere.startCrawl,
+    // me.startCrawl,
+    // ie.startCrawl,
+    // ship.startCrawl,
+    // cee.startCrawl,
+    // architecture.startCrawl,
+    // mse.startCrawl,
+    // math.startCrawl,
+    stat.startCrawl,
 ];
